requirements.txt
beautifulsoup4==4.9.3
certifi==2021.5.30
chardet==4.0.0
colorama==0.4.4
cycler==0.10.0
idna==2.10
ImageHash==4.2.0
imutils==0.5.4
kiwisolver==1.3.1
lxml==4.6.3
matplotlib==3.4.2
numpy==1.21.0
opencv-python==4.5.2.54
Pillow==8.3.1
prompt-toolkit==1.0.15
pyparsing==2.4.7
python-dateutil==2.8.1
PyWavelets==1.1.1
requests==2.25.1
scipy==1.7.0
SelectMenu==1.0.0b2
six==1.16.0
soupsieve==2.2.1
urllib3==1.26.6
wcwidth==0.2.5
